Change acquire to take a mutable throttle.
This is part of the concurrency story. Use of Cell felt like it was
going to create surprising situations when used across multiple threads,
so by making calls to acquire/2 require a mutable reference to the
throttle in question it means that borrowck will enforce thread safety,
without sacrificing the minimalist API for single-threaded use cases.